+++ go.mod
@@ -2,7 +2,6 @@ module timezone-converter
 
 go 1.18
 
-require (
-	github.com/gorilla/mux v1.8.0
-	github.com/mattn/go-sqlite3 v1.14.17
-)
+require github.com/gorilla/mux v1.8.0
+
+require github.com/mattn/go-sqlite3 v1.14.17
