+++ main_test.go
@@ -1,15 +1,30 @@
 package main
 
 import (
-	"net/http"
-	"net/http/httptest"
-	"strings"
 	"testing"
 )
 
-func TestAddTimezone(t *testing.T) {
-	w := httptest.NewRecorder()
-	r := httptest.NewRequest(http.MethodPost, "/timezone", strings.NewReader("Test/Timezone"))
+// func TestAddTimezone(t *testing.T) {
+// w := httptest.NewRecorder()
+// r := httptest.NewRequest(http.MethodPost, "/timezone", strings.NewReader("Test/Timezone"))
 
-	addTimezone(w, r)
+// 	addTimezone(w, r)
+// }
+
+func TestRegister(t *testing.T) {
+	// read input from payload
+	// add default timeslots
+	// insert into db
+
+	// t.Run("register user", func(t *testing.T) {
+	// 	var jsonStr = []byte(`{username: "Test user", password: "password"}`)
+	// 	request := httptest.NewRequest(http.MethodPost, "/register", bytes.NewBuffer(jsonStr))
+	// 	response := httptest.NewRecorder()
+
+	// 	register(response, request)
+
+	// 	fmt.Println(response.Body)
+
+	// })
+
 }
